+++ internal/model/review.go
@@ -10,7 +10,7 @@ import (
 type Review struct {
 	ID           uint           `json:"review_id"`
 	UserId       uuid.UUID      `json:"user_id" gorm:"type:char(36)"`
-	ExpressionId uint           `json:"expression_id"`
+	ExpressionId uint64         `json:"expression_id"`
 	Content      string         `json:"content"`
 	DeletedAt    gorm.DeletedAt `gorm:"index"`
 	Time         time.Time      `gorm:"type:timestamp;default:CURRENT_TIMESTAMP" json:"time"`
